+++ e2e/tests/test-utils.js
@@ -15,13 +15,13 @@ export async function open_alert_page_and_filter(page, filter) {
     await page.waitForTimeout(500);
 }
 
-export async function fill_field_condition(page, input, option, value) {
-    await page.getByRole('button', { name: 'add_circle' }).first().click();
+export async function fill_field_condition(page, input, option, value, nth= 0) {
+    await page.getByRole('button', { name: 'add_circle' }).nth(nth).click();
     await page.waitForTimeout(200);
-    await page.locator('#field-input').fill(input);
+    await page.locator('#field-input').nth(nth).fill(input);
     await page.waitForTimeout(200);
-    await page.getByText('arrow_drop_down').nth(2).click();
+    await page.getByText('arrow_drop_down').nth(nth * 3 + 2).click();
     await page.getByRole('option', { name: option }).click();
-    await page.locator('#value').fill(value);
+    await page.locator('#value').nth(nth).fill(value);
     await page.waitForTimeout(200);
 }